2D_fitting/requirements.txt
numpy
arrgh
matplotlib
torch
einops
scikit-image
pillow
scipy
opencv-python
imageio
fire